# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Soiree API
  version: 0.1.0
  description: |
    Soiree — Robocon test-run video manager HTTP API.

    Phase 1 の主要リソース（masters / sessions / videos / runs / markers / search）を契約として定義する。
    実装は spec-devflow.md §4〜§9 の順に handler / service / UI を縦断で積む。

    ## 共通方針

    - JSON フィールド名は camelCase
    - ID は UUID 文字列
    - 時刻は ISO 8601 (RFC 3339) タイムスタンプ
    - 動画上の時間は秒単位（int = 整数秒、float = 小数秒）
    - リスト系は `{ data: [...], pagination: {...} }` のエンベロープ
    - ページングは cursor ベース (`?cursor=...&limit=...`)
    - エラーは `ErrorResponse` で統一
servers:
  - url: http://localhost:8080
    description: Local development
  - url: /
    description: Same-origin (production / docker)
security:
  - xUserId: []
...
paths:
...
  /teams/{teamId}:
    parameters:
      - $ref: "#/components/parameters/TeamId"
    get:
      tags: [teams]
      operationId: getTeam
      summary: チーム取得
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [teams]
      operationId: updateTeam
      summary: チーム更新
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTeamRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [teams]
      operationId: deleteTeam
      summary: チーム削除
      responses:
        "204":
          description: Deleted
        "404":
          $ref: "#/components/responses/NotFound"
...
  /tags:
    get:
      tags: [tags]
      operationId: listTags
      summary: タグ一覧
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TagList"
    post:
      tags: [tags]
      operationId: createTag
      summary: タグ作成
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTagRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tag"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
  /tags/{tagId}:
    parameters:
      - $ref: "#/components/parameters/TagId"
    get:
      tags: [tags]
      operationId: getTag
      summary: タグ取得
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tag"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [tags]
      operationId: updateTag
      summary: タグ更新
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTagRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tag"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
    delete:
      tags: [tags]
      operationId: deleteTag
      summary: タグ削除
      responses:
        "204":
          description: Deleted
        "404":
          $ref: "#/components/responses/NotFound"

  # ---------- Sessions ----------
...
  /tournaments/{tournamentId}/scouting-notes:
    parameters:
      - $ref: "#/components/parameters/TournamentId"
    get:
      tags: [scouting-notes]
      operationId: listScoutingNotesByTournament
      summary: 大会配下の ScoutingNote 一覧
      description: |
        (tournamentId, teamId) ごとに最大1件。Hocuspocus が ydoc_state を更新する。
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScoutingNoteList"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /scouting-notes/{noteId}:
    parameters:
      - $ref: "#/components/parameters/ScoutingNoteId"
    get:
      tags: [scouting-notes]
      operationId: getScoutingNote
      summary: ScoutingNote 取得 (メタデータ + plain_text のみ)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScoutingNote"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [scouting-notes]
      operationId: deleteScoutingNote
      summary: ScoutingNote 削除
      responses:
        "204":
          description: Deleted
        "404":
          $ref: "#/components/responses/NotFound"
...
  /search/runs:
    get:
      tags: [search]
      operationId: searchRuns
      summary: Run の横断検索
      description: |
        spec-devflow.md §9 で扱う Phase 1 の検索条件。
        期間 / Robot / Scenario / tags / Marker category / memo の組合せで Run を絞り込む。
      parameters:
        - name: tournamentId
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: from
          in: query
          required: false
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          required: false
          schema:
            type: string
            format: date-time
        - name: robotId
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: scenarioId
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: tagIds
          in: query
          description: タグ ID（カンマ区切り、AND 条件）
          required: false
          explode: false
          schema:
            type: array
            items:
              type: string
              format: uuid
        - name: markerCategories
          in: query
          description: 含まれている Marker のカテゴリ（カンマ区切り、OR 条件）
          required: false
          explode: false
          schema:
            type: array
            items:
              $ref: "#/components/schemas/MarkerCategory"
        - name: q
          in: query
          description: memo の全文検索（pg_trgm）
          required: false
          schema:
            type: string
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunList"

  # ---------- Tournaments (Phase 2) ----------
...
  /tournaments/{tournamentId}:
    parameters:
      - $ref: "#/components/parameters/TournamentId"
    get:
      tags: [tournaments]
      operationId: getTournament
      summary: Tournament 取得
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tournament"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [tournaments]
      operationId: updateTournament
      summary: Tournament 更新
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTournamentRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tournament"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [tournaments]
      operationId: deleteTournament
      summary: Tournament 削除
      responses:
        "204":
          description: Deleted
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    xUserId:
      type: apiKey
      in: header
      name: X-User-Id
      description: |
        現在ユーザーの UUID。セルフホスト・全員同権限の前提のため
        Phase 1 はクライアントが選択ユーザーを header で送る簡易方式。
        将来正式な認証を入れたら securityScheme を差し替える。
  parameters:
    Cursor:
      name: cursor
      in: query
      required: false
      description: 前回レスポンスの `pagination.nextCursor`
      schema:
        type: string
    Limit:
      name: limit
      in: query
      required: false
      description: 1 レスポンスあたり最大件数（1〜200、既定 50）
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
...
    TeamId:
      name: teamId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    TagId:
      name: tagId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    TournamentId:
      name: tournamentId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    ScoutingNoteId:
      name: noteId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    BadRequest:
      description: リクエストが不正
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: リソースが存在しない
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    Conflict:
      description: 競合（重複や状態不一致）
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: 機械可読なエラーコード
        message:
          type: string
        details:
          type: object
          additionalProperties: true
    Pagination:
      type: object
      required: [hasMore]
      properties:
        hasMore:
          type: boolean
        nextCursor:
          type: [string, "null"]
          description: 次ページを取得するための cursor

    # ---------- Enums ----------
...
    MarkerCategory:
      type: string
      enum: [success, failure, note]
      description: Marker のカテゴリ。spec の `success | failure | note` を採用

    # ---------- User ----------
...
    Team:
      type: object
      required: [id, name, isOwn, createdAt]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        isOwn:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    UpdateTeamRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        isOwn:
          type: boolean
...
    Tag:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        color:
          type: [string, "null"]
        createdAt:
          type: string
          format: date-time
    CreateTagRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        color:
          type: [string, "null"]
    UpdateTagRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        color:
          type: [string, "null"]
    TagList:
      type: object
      required: [data, pagination]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Tag"
        pagination:
          $ref: "#/components/schemas/Pagination"

    # ---------- Session ----------
...
    Run:
      type: object
      required: [id, tournamentId, sessionId, teamId, robotId, scenarioId, startedAt, endedAt, durationSec, memo, createdAt]
      description: |
        `endedAt` はサーバー側で `startedAt + durationSec` から計算したレスポンス専用の値で、更新時は `startedAt` と `durationSec` を渡す。
      properties:
        id:
          type: string
          format: uuid
        tournamentId:
          type: string
          format: uuid
        sessionId:
          type: string
          format: uuid
        teamId:
          type: string
          format: uuid
        robotId:
          type: string
          format: uuid
        scenarioId:
          type: string
          format: uuid
        matchId:
          type: [string, "null"]
          format: uuid
        startedAt:
          type: string
          format: date-time
        endedAt:
          type: string
          format: date-time
        durationSec:
          type: integer
          minimum: 0
          description: Run のタイムライン長（秒）。プレイヤーはこの値を上限として表示する
        score:
          type: [number, "null"]
        memo:
          type: string
        videos:
          type: array
          description: getRun のみで埋まる。listRuns では省略
          items:
            $ref: "#/components/schemas/RunVideo"
        tagIds:
          type: array
          items:
            type: string
            format: uuid
        createdAt:
          type: string
          format: date-time
...
    RunList:
      type: object
      required: [data, pagination]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Run"
        pagination:
          $ref: "#/components/schemas/Pagination"

    # ---------- RunVideo ----------
    RunVideo:
      type: object
      required: [id, runId, videoId, videoOffsetStartSec, videoOffsetEndSec, runOffsetSec, angleLabel]
      properties:
        id:
          type: string
          format: uuid
        runId:
          type: string
          format: uuid
        videoId:
          type: string
          format: uuid
        videoOffsetStartSec:
          type: integer
          description: Source video 側で「Run の中で見せる」区間の開始秒
        videoOffsetEndSec:
          type: integer
          description: Source video 側で「Run の中で見せる」区間の終了秒
        runOffsetSec:
          type: integer
          minimum: 0
          description: Run タイムライン上でこの動画が登場する秒。0 なら Run 開始と同時
        angleLabel:
          type: string
          description: 例 "正面" / "コート横" / "drone"
...
    Tournament:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        startDate:
          type: [string, "null"]
          format: date
        endDate:
          type: [string, "null"]
          format: date
        createdAt:
          type: string
          format: date-time
...
    UpdateTournamentRequest:
      type: object
      properties:
        name:
          type: string
        startDate:
          type: [string, "null"]
          format: date
        endDate:
          type: [string, "null"]
          format: date
...
    ScoutingNote:
      type: object
      required: [id, tournamentId, teamId, plainText, updatedAt, createdAt]
      properties:
        id:
          type: string
          format: uuid
        tournamentId:
          type: string
          format: uuid
        teamId:
          type: string
          format: uuid
        plainText:
          type: string
          description: Hocuspocus が yjs ドキュメントから派生して書き込む検索用テキスト
        updatedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
    ScoutingNoteList:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/ScoutingNote"